Add APIV1ChunkedPackageCache
PackageListing list for Lethal Company has grown so large, that the
download timeouts for people with slow connections. Furthermore sooner
or later we will go over the limit of data we can pass to mod manager's
plugin layer in one go.

To solve these issues, the package list will be downloaded and
processed in chunks. API will provide an index file which contains URLs
to the chunks of package info. All data is served in gzip compressed
blob files so CDN can cache it efficiently.