Add periodic task for updating APIV1ChunkedPackageCache
New cache files are created at the top of the hour and the stale cache
is dropped. Note that currently the latter is not implemented, since we
have to figure out how to safely delete blob files from main and mirror
storages without ending up in situations where the file is deleted from
one but still exists in the other due to some error.